+++ example/swagger/codegen-long.js
@@ -3,7 +3,7 @@ const { codegen } = require('../../dist/index.js')
 
 codegen({
   methodNameMode: 'path',
-  source: require('../swagger.json'),
+  source: require('../api-docs.json'),
   // remoteUrl: 'http://localhost:44307/swagger/v1/swagger.json',
   outputDir: './swagger/services',
   strictNullChecks: false,