+++ src/templates/serviceHeader.ts
@@ -1,21 +1,38 @@
 import * as fs from 'fs'
 import * as path from 'path'
-import { ISwaggerOptions } from "../baseInterfaces";
-import { abpGenericTypeDefinition, universalGenericTypeDefinition } from './genericTypeDefinitionTemplate';
-import { trimString } from '../utils';
+import { ISwaggerOptions } from '../baseInterfaces'
+import { abpGenericTypeDefinition, universalGenericTypeDefinition } from './genericTypeDefinitionTemplate'
 
 export function serviceHeader(options: ISwaggerOptions) {
   const classTransformerImport = options.useClassTransformer
     ? `import { Expose, Transform, Type, plainToClass } from 'class-transformer';
-  ` : '';
+  `
+    : ''
   return `/** Generate by swagger-axios-codegen */
   /* eslint-disable */
   // @ts-nocheck
-  import axiosStatic, { AxiosInstance, AxiosRequestConfig } from 'axios';
+  import axiosStatic, { type AxiosInstance, type AxiosRequestConfig } from 'axios';
 
   ${classTransformerImport}
 
   export interface IRequestOptions extends AxiosRequestConfig {
+    /**
+     * show loading status
+     */
+    loading?: boolean;
+    /**
+     * display error message
+     */
+    showError?: boolean;
+    /**
+     * data security, extended fields are encrypted using the specified algorithm
+     */
+    security?: Record<string, 'md5' | 'sha1' | 'aes' | 'des'>;
+    /**
+     * indicates whether Authorization credentials are required for the request
+     * @default true
+     */
+    withAuthorization?: boolean;
   }
 
   export interface IRequestConfig {
@@ -28,11 +45,14 @@ export function serviceHeader(options: ISwaggerOptions) {
 
   // Add options interface
   export interface ServiceOptions {
-    axios?: AxiosInstance,
+    axios?: AxiosInstance;
+    /** only in axios interceptor config*/
+    loading:boolean;
+    showError:boolean;
   }
 
   ${requestHeader()}
-  `;
+  `
 }
 
 export function disableLint() {
@@ -40,16 +60,33 @@ export function disableLint() {
   // @ts-nocheck
 /* eslint-disable */
   
-`}
-
+`
+}
 
 export function customerServiceHeader(options: ISwaggerOptions) {
-
   return `/** Generate by swagger-axios-codegen */
   // @ts-nocheck
   /* eslint-disable */
-  export interface IRequestOptions {
-    headers?: any;
+  import axiosStatic, { type AxiosInstance, type AxiosRequestConfig } from 'axios';
+
+  export interface IRequestOptions extends AxiosRequestConfig {
+    /**
+     * show loading status
+     */
+    loading?: boolean;
+    /**
+     * display error message
+     */
+    showError?: boolean;
+    /**
+     * data security, extended fields are encrypted using the specified algorithm
+     */
+    security?: Record<string, 'md5' | 'sha1' | 'aes' | 'des'>;
+    /**
+     * indicates whether Authorization credentials are required for the request
+     * @default true
+     */
+    withAuthorization?: boolean;
   }
 
   export interface IRequestPromise<T=any> extends Promise<IRequestResponse<T>> {}
@@ -79,7 +116,10 @@ export function customerServiceHeader(options: ISwaggerOptions) {
 
   // Add options interface
   export interface ServiceOptions {
-    axios?: IRequestInstance,
+    axios?: IRequestInstance;
+    /** only in axios interceptor config*/
+    loading:boolean;
+    showError:boolean;
   }
 
   ${requestHeader()}
@@ -109,7 +149,7 @@ function requestHeader() {
   }
   
   export function getConfigs(method: string, contentType: string, url: string,options: any):IRequestConfig {
-    const configs: IRequestConfig = { ...options, method, url };
+    const configs: IRequestConfig = { loading:serviceOptions.loading, showError:serviceOptions.showError, ...options, method, url };
     configs.headers = {
       ...options.headers,
       'Content-Type': contentType,